requirements.txt
# SlideSync Requirements
# Python 3.10+ required

# Screen Capture
mss>=9.0.1
Pillow>=10.0.0

# Image Processing & Deduplication
imagehash>=4.3.1

# Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
python-jose[cryptography]>=3.3.0

# GUI
customtkinter>=5.2.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# LLM Integration (Optional)
# openai>=1.3.0
# anthropic>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Logging
coloredlogs>=15.0.1

# Utilities
python-dateutil>=2.8.2

# Admin Panel & Authentication
qrcode[pil]>=7.4.2

# Wayland Portal Support (Linux)
PyGObject>=3.42.0
pydbus>=0.6.0
